Extract reusable CoreData heleprs into a separate file
